fir_userinteraction/requirements.txt
django-bootstrap-form
django-ipware
django-markdownx
celery
python-ldap
pytz
python-dateutil
oauth2_sso
